Refactor: Move GameRecord to models package
