+++ app/src/main/java/com/example/infiltrados/models/GameRecord.kt
@@ -1,0 +1,17 @@
+package com.example.infiltrados.models
+
+import com.example.infiltrados.services.MultiplayerPhase
+import com.google.gson.annotations.SerializedName
+import kotlinx.serialization.Serializable
+
+
+@Serializable
+data class GameRecord(
+    @SerializedName("\$id")
+    val id: String,
+    val players: List<Player>,
+    val phase: MultiplayerPhase,
+    val word1: String,
+    val word2: String,
+    val state: String
+)
